requirements.txt
flask
openai
whisper
pillow
pyautogui
opencv-python-headless
